requirements.txt
# requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
alembic==1.13.1
flask-migrate==4.0.5